Add toggle to enable verbose logs
Closes #189.
