+++ BTCPayApp.Server/appsettings.Development.json
@@ -2,7 +2,7 @@
   "DetailedErrors": true,
   "Logging": {
     "LogLevel": {
-      "Default": "Information",
+      "Default": "Warning",
       "Microsoft.AspNetCore": "Warning",
       "System.Net.Http": "Warning",
       "LDK": "Trace",
@@ -10,6 +10,20 @@
       "LDK.lightning::routing::gossip": "Information",
       "LDK.BTCPayApp.Core.LDK.LDKPeerHandler": "Information",
       "LDK.lightning_background_processor": "Information"
+    }
+  },
+  "Serilog": {
+    "MinimumLevel": {
+      "Default": "Warning",
+      "Override": {
+        "Microsoft.AspNetCore": "Warning",
+        "System.Net.Http": "Warning",
+        "LDK": "Trace",
+        "LDK.lightning::ln::peer_handler": "Information",
+        "LDK.lightning::routing::gossip": "Information",
+        "LDK.BTCPayApp.Core.LDK.LDKPeerHandler": "Information",
+        "LDK.lightning_background_processor": "Information"
+      }
     }
   }
 }
